refactor: move dataLogging classes to main package and update imports
- Moved classes from 'dataLogging' to 'com.skeletonarmy.marrow'
- Renamed `createTar` to `createTarFile` for clarity
- Added overloaded `createTarFile` method for directory input
- Updated `NotDirectoryException` to extend IOException
- Adjusted all relevant import statements accordingly
